refactor: move interfaces to a separate submodule
Moved all interface definitions
