+++ src/bitrix24_client/interfaces/response_validator.py
