+++ src/bitrix24_client/retry_strategies/linear.py
@@ -1,4 +1,4 @@
-from .interfaces import RetryStrategyI
+from ..interfaces import RetryStrategyI
 
 
 class LinearRetryStrategyI(RetryStrategyI):
