+++ src/bitrix24_client/retry_strategies/logarithmic.py
@@ -1,6 +1,6 @@
 from math import log1p
 
-from .interfaces import RetryStrategyI
+from ..interfaces import RetryStrategyI
 
 
 class LogarithmicRetryStrategyI(RetryStrategyI):
